If statement
The if statment is going to see if it should print a # or a space. This is true for positions in the row that aren't the first or last hashtag. If the if statment is true, a space will be printed using (" ", end=""). If its not, the else statment will print using print("#", end=""). This should print a # at the end of the row.