name error
For the last part I created a loop to show the prime numbers 1 thorugh 100. i used the is_prime function to see if the number is prime. If it returens true, the number is printed. But I have an error with some of my statements that I'm going to try and fix