fixed code
Okay I just got really confused and took like 30 minutes to figure this out but I looked up some things and got it. In order to generate a response with a list of numbers, i opened a textfile and put the numbers that are in my output in that file all on their own line. Now python can find this file and a list of numbers is generated