Enter sales for each day of week
Now I created a for loop to use over the days of the week.  I tell the user to enter an amount for each day using the input function. Then I used float since its not going to be whole numbers. I used the append method to append it to the 'sales'